+++ src/main/java/com/flansmod/common/network/PacketDriveableDamage.java
@@ -16,25 +16,25 @@ import com.flansmod.common.driveables.EnumDriveablePart;
 public class PacketDriveableDamage extends PacketBase
 {
 	public int entityId;
-	public short[] health;
+	public int[] health;
 	public boolean[] onFire;
 	
 	public PacketDriveableDamage()
 	{
-		health = new short[EnumDriveablePart.values().length];
+		health = new int[EnumDriveablePart.values().length];
 		onFire = new boolean[EnumDriveablePart.values().length];
 	}
 	
 	public PacketDriveableDamage(EntityDriveable driveable)
 	{
 		entityId = driveable.getEntityId();
-		health = new short[EnumDriveablePart.values().length];
+		health = new int[EnumDriveablePart.values().length];
 		onFire = new boolean[EnumDriveablePart.values().length];
 		for(int i = 0; i < EnumDriveablePart.values().length; i++)
 		{
 			EnumDriveablePart ep = EnumDriveablePart.values()[i];
 			DriveablePart part = driveable.getDriveableData().parts.get(ep);
-			health[i] = (short)part.health;
+			health[i] = part.health;
 			onFire[i] = part.onFire;
 		}
 	}
@@ -45,7 +45,7 @@ public class PacketDriveableDamage extends PacketBase
 		data.writeInt(entityId);
 		for(int i = 0; i < EnumDriveablePart.values().length; i++)
 		{
-			data.writeShort(health[i]);
+			data.writeInt(health[i]);
 			data.writeBoolean(onFire[i]);
 		}
 	}
@@ -56,7 +56,7 @@ public class PacketDriveableDamage extends PacketBase
 		entityId = data.readInt();
 		for(int i = 0; i < EnumDriveablePart.values().length; i++)
 		{
-			health[i] = data.readShort();
+			health[i] = data.readInt();
 			onFire[i] = data.readBoolean();
 		}
 	}
